quick_test.py
#!/usr/bin/env python3
"""
Quick test to verify the tensor creation fix works.
This creates a simple test scenario to check if our fix resolves the padding issue.
"""

import sys
import tempfile
from pathlib import Path
import numpy as np
from PIL import Image

def create_simple_test_frames(num_frames=8):
    """Create simple test frames."""
    frames = []
    for i in range(num_frames):
        # Create a 224x224 RGB image with different colors per frame
        img_array = np.full((224, 224, 3), fill_value=(i * 30) % 255, dtype=np.uint8)
        frame = Image.fromarray(img_array, 'RGB')
        frames.append(frame)
    return frames

def test_tensor_creation():
    """Test the tensor creation with our fix."""
    print("🧪 Testing Tensor Creation Fix")
    print("=" * 40)

    try:
        # Import required modules
        from transformers import AutoImageProcessor
        import torch
        print("✅ Imports successful")

        # Load processor
        processor = AutoImageProcessor.from_pretrained("facebook/timesformer-base-finetuned-k400")
        print("✅ Processor loaded")

        # Create test frames
        frames = create_simple_test_frames(8)
        print(f"✅ Created {len(frames)} test frames")

        # Test our fix approach
        try:
            inputs = processor(images=frames, return_tensors="pt", padding=True)
            print(f"✅ Tensor created successfully!")
            print(f"   Shape: {inputs['pixel_values'].shape}")
            print(f"   Dtype: {inputs['pixel_values'].dtype}")
            return True

        except Exception as e:
            print(f"❌ Primary approach failed: {e}")

            # Try fallback
            try:
                inputs = processor(images=[frames], return_tensors="pt", padding=True)
                print(f"✅ Fallback approach worked!")
                print(f"   Shape: {inputs['pixel_values'].shape}")
                return True
            except Exception as e2:
                print(f"❌ Fallback also failed: {e2}")
                return False

    except Exception as e:
        print(f"❌ Test setup failed: {e}")
        return False

def test_prediction_pipeline():
    """Test the full prediction pipeline."""
    print("\n🎬 Testing Full Pipeline")
    print("=" * 40)

    try:
        from predict import predict_actions
        print("✅ Import successful")

        # Create a temporary video file (simulate with images)
        with tempfile.TemporaryDirectory() as tmp_dir:
            tmp_path = Path(tmp_dir)

            # For this test, we'll create a simple video-like structure
            # Since we can't easily create a real video, we'll test the frame processing directly

            # This would normally be called by predict_actions with a real video file
            print("⚠️  Note: Full video test requires a real video file")
            print("   The tensor fix is now in place in predict.py")

        return True

    except Exception as e:
        print(f"❌ Pipeline test failed: {e}")
        return False

if __name__ == "__main__":
    print("🔧 Quick Test Suite for Tensor Fix")
    print("=" * 50)

    # Test 1: Basic tensor creation
    test1_passed = test_tensor_creation()

    # Test 2: Pipeline integration
    test2_passed = test_prediction_pipeline()

    print("\n📊 Results:")
    print(f"   Tensor creation: {'✅ PASSED' if test1_passed else '❌ FAILED'}")
    print(f"   Pipeline check: {'✅ PASSED' if test2_passed else '❌ FAILED'}")

    if test1_passed:
        print("\n🎉 The tensor creation fix appears to be working!")
        print("   You can now try uploading a video to the Streamlit app.")
    else:
        print("\n💥 The fix may need more work. Check the error messages above.")

    print("\n💡 Next step: Run 'streamlit run app.py' and test with a real video")
